requirements.txt

### **3. MISE À JOUR DU requirements.txt** (Version finale)
```txt
# === CORE ===
streamlit==1.28.0
pandas==2.1.0

# === PDF & RAPPORTS ===
reportlab==4.0.4
fpdf2==2.7.0
matplotlib==3.7.2

# === EMAIL & ENVIRONNEMENT ===
python-dotenv==1.0.0

# === DONNÉES ===
pycountry==23.12.0
numpy==1.24.0
openpyxl==3.1.0

# === VISUALISATION ===
plotly==5.15.0

# === IMAGES ===
Pillow==10.0.1
qrcode[pil]==7.4.2